==> 1_download to csv formats.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot,rcParams
from collections import Counter
from sqlalchemy import create_engine

# extracting the data and saving it to csv format

db_host = '18.136.56.185'
username = 'dm_team1'
user_pass = 'dm_team1123#'
db_name = 'project_banking'
conn = create_engine('mysql+pymysql://'+username+':'+user_pass+'@'+db_host+'/'+db_name)

query ='show tables'

data=pd.read_sql(query,conn)
data.head()

query='select *from Cust_Account'


data=pd.read_sql(query,conn)
data.head()

data.to_csv('Cust_Account.csv',index=False)

query='select * from Cust_Demographics'

data=pd.read_sql(query,conn)
data.head()

data.to_csv('Cust_Demographics.csv',index=False)

query= 'select * from Cust_Enquiry'

data=pd.read_sql(query,conn)

data.head()

data.to_csv('Cust_Enquiry',index=False)
