==> 2_Basic data convertion.ipynb <==
import pandas as pd
import numpy as np


# Enquiry File Transformation

data=pd.read_csv('Cust_Enquiry')
data.head()

data.columns

data1=data.loc[:,['customer_no','enq_amt']]
data1.head()

Cust_group=data1.groupby(['customer_no'])
Cust_group.head()

df_amt=pd.DataFrame(Cust_group.sum()).reset_index()
df_amt.columns = ['customer_no','total_enq_amt']
df_amt.head()
#df_amt.shape

data1_count=pd.DataFrame(Cust_group.count()).reset_index()
data1_count.columns = ['customer_no','total_enq_count']
data1_count.head()

data_final=pd.merge(data1_count,df_amt,on='customer_no',how='inner')
data_final.head()

data_final.to_csv('Cust_enq_final.csv',index=False)

# Account file transformation

data_acnt=pd.read_csv('Cust_Account.csv')
data_acnt.head()

ColList = ['customer_no','high_credit_amt', 'cur_balance_amt', 'amt_past_due', 'creditlimit','cashlimit']
data_acnt=data_acnt.loc[:,ColList]
data_acnt.head()

data_acnt.fillna(0,inplace=True)
data_acnt.head()

Cust_group_accnt=data_acnt.groupby(['customer_no'])
Cust_group_accnt.head()

data_acnt=pd.DataFrame(Cust_group_accnt.sum()).reset_index()

data_acnt.to_csv('Cust_Account_Final.csv',index=False)

# Demographic file tranformation

data_demo=pd.read_csv('Cust_Demographics.csv')
data_demo.head()

DemColList=['dt_opened', 'customer_no', 'entry_time', 'feature_1', 'feature_2',
       'feature_3', 'feature_4', 'feature_5', 'feature_6', 'feature_7',
       'feature_8', 'feature_9', 'feature_10', 'feature_11', 'feature_12',
       'feature_13', 'feature_14', 'feature_15', 'feature_16', 'feature_17',
       'feature_18', 'feature_19', 'feature_20', 'feature_21', 'feature_22',
       'feature_23', 'feature_24', 'feature_25', 'feature_26', 'feature_27',
       'feature_28', 'feature_29', 'feature_30', 'feature_31', 'feature_32',
       'feature_33', 'feature_34', 'feature_35', 'feature_36', 'feature_37',
       'feature_38', 'feature_39', 'feature_40', 'feature_41', 'feature_42',
       'feature_43', 'feature_44', 'feature_45', 'feature_46', 'feature_47',
       'feature_48', 'feature_49', 'feature_50', 'feature_51', 'feature_52',
       'feature_53', 'feature_54', 'feature_55', 'feature_56', 'feature_57',
       'feature_58', 'feature_59', 'feature_60', 'feature_61', 'feature_62',
       'feature_63', 'feature_64', 'feature_65', 'feature_66', 'feature_67',
       'feature_68', 'feature_69', 'feature_70', 'feature_71', 'feature_72',
       'feature_73', 'feature_74', 'feature_75', 'feature_76', 'feature_77',
       'feature_78', 'feature_79', 'Bad_label']

print(data_acnt.shape)
print(data_demo.shape)
print(data_final.shape)

data1_final=pd.merge(data_demo,data_acnt,on='customer_no', how='left')
data1_final.shape

data1_final=pd.merge(data1_final,data_final,on='customer_no', how='left')
data1_final.shape

data1_final.to_csv('final_data.csv',index=False)
