==> 3_Feature engineering.ipynb <==
# Feature engineering

import numpy as np
import pandas as pd


data=pd.read_csv('final_data.csv')
data.head()

data.info()

data1 = data.loc[:,['customer_no','feature_1','feature_4', 'feature_5', 
        'feature_6', 'feature_7','feature_11', 'feature_19', 
        'feature_23','feature_25', 'feature_26', 'feature_27','feature_29', 'feature_30',
        'feature_31', 'feature_32','feature_33', 'feature_34', 'feature_35', 'feature_36', 
        'feature_37', 'feature_40', 'feature_41', 'feature_42','feature_44',
        'feature_46', 'feature_48', 'feature_55', 'feature_56', 'feature_58',
        'feature_59', 'feature_60', 'feature_62','feature_64', 'feature_65', 'feature_67',
       'feature_68', 'feature_69', 'feature_71', 'feature_72','feature_76', 'feature_78', 
       'feature_79', 'Bad_label', 'high_credit_amt','cur_balance_amt', 'amt_past_due',
        'creditlimit','cashlimit', 'total_enq_amt', 'total_enq_count']]

data1.head()

data1.replace(['?','*','$',' ','  ',''],np.nan,inplace=True)

data1.isnull().sum()

data1.loc[:,['feature_36','feature_37','feature_46','feature_48']].info()

data1.loc[:,['feature_36','feature_37','feature_46','feature_48']]=data1.loc[:,['feature_36','feature_37','feature_46','feature_48']].replace(np.nan,'others')

data1.shape

from collections import Counter
Counter(data1.feature_36)

data1.drop_duplicates(inplace=True)
data1.shape

#data1.dropna(inplace=True)
#rint(data1.shape)

data1.isnull().sum()

data1.info()

data2=data1.loc[:,['customer_no','feature_1','feature_4', 'feature_5', 
        'feature_6', 'feature_7','feature_11', 'feature_19', 
        'feature_23','feature_25', 'feature_26', 'feature_27','feature_29', 'feature_30',
        'feature_31', 'feature_32','feature_33', 'feature_34', 'feature_35', 'feature_36', 
        'feature_37', 'feature_40', 'feature_41', 'feature_42','feature_44',
        'feature_46', 'feature_48', 'feature_55', 'feature_56', 'feature_58',
        'feature_59', 'feature_60', 'feature_62','feature_64', 'feature_65', 'feature_67',
       'feature_68', 'feature_69', 'feature_71', 'feature_72','feature_76', 'feature_78', 
       'feature_79', 'Bad_label', 'high_credit_amt','cur_balance_amt', 'amt_past_due',
        'creditlimit','cashlimit', 'total_enq_amt', 'total_enq_count']]

data2.head()

data2.isnull().sum()

print(data2.shape)
data2.dropna(inplace=True)
print(data2.shape)

data2.to_csv('cleansed_data_feat_engg.csv',index=False)
