==> 4_model prediction.ipynb <==
import numpy as np
import pandas as pd

data=pd.read_csv('cleansed_data_feat_engg.csv')
data.head()

data.info()

from sklearn.preprocessing import LabelEncoder
enc=LabelEncoder()

to_enc=['feature_1','feature_5','feature_11','feature_23','feature_27','feature_32',
                'feature_33','feature_36','feature_37','feature_46','feature_48','feature_58',
                'feature_59','feature_60','feature_62','feature_72','feature_79']

for i in to_enc:
    data[[i]]=enc.fit_transform(data[[i]])



data.head()

data2 = data.loc[:,['customer_no','feature_1','feature_4', 'feature_5', 
        'feature_6', 'feature_7','feature_11', 'feature_19', 
        'feature_23','feature_25', 'feature_26', 'feature_27','feature_29', 'feature_30',
        'feature_31', 'feature_32','feature_33', 'feature_34', 'feature_35', 'feature_36', 
        'feature_37', 'feature_40', 'feature_41', 'feature_42','feature_44',
        'feature_46', 'feature_48', 'feature_55', 'feature_56', 'feature_58',
        'feature_59', 'feature_60', 'feature_62','feature_64', 'feature_65', 'feature_67',
       'feature_68', 'feature_69', 'feature_71', 'feature_72','feature_76', 'feature_78', 
       'feature_79', 'high_credit_amt','cur_balance_amt', 'amt_past_due',
        'creditlimit','cashlimit', 'total_enq_amt', 'total_enq_count','Bad_label']]

data2.head()

X=data2.iloc[:,:-1]
y=data2.Bad_label

from sklearn.model_selection import train_test_split
X_train,X_test,y_train,y_test=train_test_split(X,y,random_state=10,test_size=.2)

from sklearn.ensemble import RandomForestClassifier
model=RandomForestClassifier(n_estimators=20,random_state=20)
model.fit(X_train,y_train)

y_predict=model.predict(X_test)

from sklearn.metrics import accuracy_score
accuracy_score(y_test,y_predict)

# xg boost

from xgboost import XGBClassifier
model=XGBClassifier(learning_rate=0.5,random_state=12,n_estimators=200,gamma=0.5)
model.fit(X_train,y_train)
y_predict=model.predict(X_test)

accuracy_score(y_test,y_predict)
